refactor(node/sync): extract Reconciler component (Phase 2 of #2313)
Moves the reconcile-after-divergence orchestration out of
`sync::manager::mod.rs` and into a new `sync::reconciler` module:

- `Reconciler` struct owns `Arc<dyn SyncStateAccess>` + `Arc<dyn SyncNetwork>`
  + `ContextClient` and runs `reconcile_after_divergence` /
  `reconcile_one_divergent_context` / `verify_post_reconcile_root_hash` /
  `anchor_identities_for_group`.
- `ReconcileSyncDispatch` trait carries the single `initiate_sync` call
  the reconciler makes back into the rest of the sync machinery. Passed
  per-call (not stored on the struct) to avoid a self-referential
  `SyncManager ↔ Reconciler` ownership cycle. `?Send` because
  `SyncManager::initiate_sync` is not Send-safe internally
  (`delta_store.rs` keeps a non-Send `Box<dyn Iterator>` across an
  await), and the reconciler is only ever awaited inside a single
  async task.
- Backoff math + cooldown-state helpers (`reconcile_cooldown` +
  three free fns) move alongside, keeping `crate::state`'s
  `SyncStateAccess` impl as the single consumer of the latter three.
- `SyncManager` keeps `reconcile_after_divergence` as a thin
  forwarder, so the two external callers in
  `handlers/network_event/namespace.rs` are unchanged.

`anchor_identities_for_context` stays on `SyncManager` (called by
`find_peer_with_state` for anchor-priority peer ordering, unrelated
to reconcile). `anchor_identities_for_group` is reproduced on the
reconciler — 8 lines of trivial group-store lookup, intentional small
dup to keep the reconciler self-contained.

### Diff stats

- `sync::manager::mod.rs` 4607 → 4342 LOC (-265). Three reconcile
  methods + four free fns moved.
- `sync::manager::tests.rs` -116 LOC. The `reconcile_cooldown` /
  `record_reconcile_*` test block moved to `sync::reconciler::tests`.
- `sync::reconciler.rs` +609 LOC (270 prod + module + 90 tests + doc).

### Tests

- 7 unit tests for the backoff math and cooldown-state helpers, ported
  verbatim from `sync::manager::tests`. All green.
- Reconciler-method orchestration tests (cooldown short-circuit,
  anchor-selection paths, hash-verify) need a lightweight `ContextClient`
  test fixture that today's `MockSyncStateAccess` / `MockSyncNetwork`
  pair doesn't cover. Deferred to a follow-up alongside that fixture
  work — the existing partition-window e2e workflow
  (`apps/scaffolding-e2e/workflows/member-removed-partition-window-reconcile.yml`)
  exercises the full path end-to-end in the meantime.
- `cargo test -p calimero-node --lib`: 179 passed (172 pre-existing +
  7 new).
- `cargo test -p calimero-node --test '*'`: 308 passed, no regression.
- `cargo build -p calimero-node`: clean (only pre-existing warnings).
- `cargo clippy -p calimero-node --lib --tests`: no new warnings in
  touched files.

### Behaviour preservation

All method bodies move verbatim — `tracing::*` calls, doc comments,
field references unchanged. The only line-level edit inside the moved
code is `self.initiate_sync(...)` → `dispatch.initiate_sync(...)`,
which forwards through the trait impl to the same fn.
